Refactor authentication middleware to restrict access based on user type
This commit refactors the authentication middleware in authController.js to add a new function called restrictTo. This function checks if the user's userType is included in the provided array of user types. If not, it returns a 403 Forbidden error. This change allows for more granular access control based on user types.