Refactor projectController.js to include user model in getAllProject route
This commit modifies the projectController.js file to include the user model in the getAllProject route. By using the opion in t method,de the  the user details associated with each project will be included in the JSON response. This change enhances the functionality of the getAllProject route by providing additional information about the users who created the projects.